Creating logic for saving coupons and products and their calculations

+++ src/main/java/by/rishat/ws/systemio/service/PaymentOrchestrator.java
@@ -1,10 +1,11 @@
 package by.rishat.ws.systemio.service;
 
-import by.rishat.ws.systemio.dto.customException.paymentExceptions.NoProcessorAvailable;
+import by.rishat.ws.systemio.exception.paymentExceptions.NoProcessorAvailable;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
+import java.math.BigDecimal;
 import java.util.List;
 import java.util.concurrent.ThreadLocalRandom;
 
@@ -15,7 +16,7 @@ public class PaymentOrchestrator {
 
     private final List<PaymentService> paymentServices;
 
-    public void executeRandom(double amount) {
+    public void executeRandom(BigDecimal amount) {
         if (paymentServices.isEmpty()) {
             throw new NoProcessorAvailable("Не найден платежный процессор");
         }
