+++ src/main/java/by/rishat/ws/systemio/service/PaymentService.java
@@ -1,7 +1,9 @@
 package by.rishat.ws.systemio.service;
 
+import java.math.BigDecimal;
+
 //Реализуем Адаптер, для "Учитывайте возможность добавления новых платежных процессоров."
 public interface PaymentService {
     String getName();
-    void process(double amount);
+    void process(BigDecimal amount);
 }
