+++ src/main/java/by/rishat/ws/systemio/service/payment/StripePaymentService.java
@@ -1,9 +1,11 @@
 package by.rishat.ws.systemio.service.payment;
 
-import by.rishat.ws.systemio.dto.customException.paymentExceptions.StripePaymentException;
+import by.rishat.ws.systemio.exception.paymentExceptions.StripePaymentException;
 import by.rishat.ws.systemio.service.PaymentService;
 import by.rishat.ws.systemio.service.payment.mock.StripePaymentProcessor;
 import org.springframework.stereotype.Service;
+
+import java.math.BigDecimal;
 
 @Service
 public class StripePaymentService implements PaymentService {
@@ -15,8 +17,8 @@ public class StripePaymentService implements PaymentService {
     }
 
     @Override
-    public void process(double amount) {
-        boolean success = stripePaymentProcessor.pay((float) amount);
+    public void process(BigDecimal amount) {
+        boolean success = stripePaymentProcessor.pay((amount.floatValue()));
         if (!success) {
             throw new StripePaymentException("Цена меньше 100");
         }
